size of lexer int made matching to the parser's variables.

+++ 01/lexer.cpp
@@ -16,12 +16,12 @@ TokenType Lexer::getNext() {
         m_lastValue = letter - '0';
         char letter2;
         while (m_input.size() >= m_charIdx && std::isdigit(letter2 = m_input[m_charIdx])) {
-            if (m_lastValue < (std::numeric_limits<unsigned long long>::max() - 9) / 10) {
+            if (m_lastValue < (std::numeric_limits<unsigned long>::max() - 9) / 10) {
                 m_lastValue = 10 * m_lastValue + (letter2 - '0');
             } else {
-                if (m_lastValue <= std::numeric_limits<unsigned long long>::max() / 10) {
+                if (m_lastValue <= std::numeric_limits<unsigned long>::max() / 10) {
                         m_lastValue *= 10;
-                    if (m_lastValue <= std::numeric_limits<unsigned long long>::max() - (letter2 - '0')) {
+                    if (m_lastValue <= std::numeric_limits<unsigned long>::max() - (letter2 - '0')) {
                         m_lastValue += letter2 - '0';
                     } else {
                         throw ErrorCode::INPUT_OVERFLOW;
